# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JobQuest API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List user's applications
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: sortBy
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"

    post:
      operationId: createApplication
      tags: [applications]
      summary: Create an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a single application
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete application
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── RESUMES ────────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{id}/suspend:
    patch:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Suspended
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, company, location, source, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        companyLogo:
          type: ["string", "null"]
        location:
          type: string
        workMode:
          type: string
          enum: [remote, hybrid, onsite]
        experienceLevel:
          type: ["string", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        salaryCurrency:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        source:
          type: string
        applyUrl:
          type: ["string", "null"]
        isNew:
          type: boolean
        isHot:
          type: boolean
        matchScore:
          type: ["integer", "null"]
        postedAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        isSaved:
          type: boolean
...
    Application:
      type: object
      required: [id, userId, company, role, status, appliedAt, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        jobId:
          type: ["integer", "null"]
        company:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [saved, applied, under_review, pending, interview_scheduled, in_process, rejected, ghosted, offer_received, accepted, declined]
        appliedAt:
          type: ["string", "null"]
        resumeId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        interviewDate:
          type: ["string", "null"]
        recruiterName:
          type: ["string", "null"]
        recruiterEmail:
          type: ["string", "null"]
        salaryOffered:
          type: ["integer", "null"]
        source:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        job:
          $ref: "#/components/schemas/Job"
    ApplicationInput:
      type: object
      required: [company, role, status]
      properties:
        jobId:
          type: ["integer", "null"]
        company:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [saved, applied, under_review, pending, interview_scheduled, in_process, rejected, ghosted, offer_received, accepted, declined]
        appliedAt:
          type: ["string", "null"]
        resumeId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [saved, applied, under_review, pending, interview_scheduled, in_process, rejected, ghosted, offer_received, accepted, declined]
        notes:
          type: ["string", "null"]
        interviewDate:
          type: ["string", "null"]
        recruiterName:
          type: ["string", "null"]
        recruiterEmail:
          type: ["string", "null"]
        salaryOffered:
          type: ["integer", "null"]
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
        applicationCount:
          type: integer
        resumeCount:
          type: integer
        createdAt:
          type: string
    AdminUserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer